backend Web Viewer/sportmind_viewer/requirements.txt
Flask>=3.0
boto3>=1.34
python-dotenv>=1.0
Flask-Cors>=4.0
requests>=2.32
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

